Add function to summarise vaccines data
This function adds the data_vaccines.csv file to the processed data dir.
The generated data is in this format:
date | mtl_doses | qc_doses | mtl_perc | qc_perc

The approx % vaccinated calculation is based on the somewhat inaccurate assumption that 2 doses = 1 person vaccinated.
This estimate should be closer to the truth once a larger % of the population is vaccinated, as the ratio of (ppl waiting for a 2nd dose / ppl having received 2 doses) decreases.